feat: integrate microservices with Consul discovery

+++ accounts/src/main/java/theo_viti/com/example/accounts/model/Customer.java
@@ -2,9 +2,6 @@ package theo_viti.com.example.accounts.model;
 
 import jakarta.persistence.*;
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
 import java.time.LocalDate;
 
 @Entity
@@ -14,7 +11,7 @@ public class Customer {
       @Id
       @GeneratedValue(strategy = GenerationType.AUTO)
       @Column(name = "customer_id")
-      private int customerId;
+      private Long customerId;
       private String name;
       private String email;
